After thought, removed Triploon from this branch
It was outside the scope of the revised Elemental Saw plans, and I originally only added it to this branch because I was still under the impression that it was going into Elemental Saw.
I have made a local branch with all of Triploon's code in case we decide we still to revive it in the future.