Fixed Cal Clone phase transitions getting offset by dealing high damage in a single hit
BEFORE: newAI[0] starts as her max health, and is set to her current health when a transition attack starts. This means dealing high damage on the hit taking into a transition attack causes the succeeding transitions to be offset.
AFTER: newAI[0] starts as 1, and has 0.3 subtracted from it when a transition attack starts. Since it no longer relies on her current health to determine suceeding phases, transitions are no longer offset.